+++ Assets/Scripts/Utils/ColorUtils.cs
@@ -7,7 +7,7 @@ public static class ColorUtils
 {
    private static Color[] clientColors = { Color.white, Color.blue, Color.green, Color.red, Color.yellow, Color.gray, Color.black, Color.cyan, Color.magenta, };
 
-   public static Color GetColorForClient(int clientID) => clientColors[clientID];
+   public static Color GetColorForClient(ulong clientID) => clientColors[clientID];
 
    public static Color GetColorWithEnum(BulletColor bulletColor)
    {
